refactor: Move registries to a new crate
