+++ crates/registries/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "registries"
+version = "0.1.0"
+edition = "2024"
+build = "build/main.rs"
+
+[build-dependencies]
+minecraft_protocol = { path = "../minecraft_protocol" }
+registries_data = { path = "../registries_data" }
+anyhow = { workspace = true }
+proc-macro2 = { workspace = true }
+quote = { workspace = true }
+serde_json = { workspace = true }
+walkdir = { workspace = true }
+
+[dependencies]
+minecraft_protocol = { path = "../minecraft_protocol" }
+registries_data = { path = "../registries_data" }
