+++ crates/registries_data/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod grouped_registries;
+pub mod registry_format;
