+++ crates/registries_data/src/registry_format.rs
