--- pico_limbo/src/registries/mod.rs
@@ -1,1 +1,0 @@
-pub mod get_registries;
